fastapi_movietheatre/fastapi-solution/tests/functional/requirements.txt
aiohttp==3.11.2
elasticsearch==8.16.0
pydantic-settings==2.6.1
pytest==8.3.3
pytest-asyncio==0.24.0
redis==5.0.4
pyyaml==6.0.2

python-jose==3.3.0